requirements.txt
requests
numpy
tqdm
python-dotenv
rank_bm25
sentence-transformers
psycopg[binary]
beautifulsoup4
lxml
pandas
openpyxl
xlrd
PyPDF2
